moved mock platform plus renaming
